register IS_PUBLISHED on our own database connection
Ideally the plugin shouldn't open it's own connection, but rely on the
main struct connection. But since our helper inherits from the struct
helper it is treated as a different class (not as the same as struct).
So we need to ensure that our function is registered on the correct
connection.

Why this has worked before is unclear. With the devel branch of the
sqlite plugin it fails.

In the future this should be refactored correctly.
